requirements.txt
flask~=3.1.0
cerebras_cloud_sdk
python-dotenv~=1.1.0
prompt_toolkit~=3.0.48
Authlib~=1.5.2